# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Memento booking and availability API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  contacted: 'contacted',
  confirmed: 'confirmed',
  declined: 'declined',
  cancelled: 'cancelled',
} as const;

export type CustomerType = typeof CustomerType[keyof typeof CustomerType];


export const CustomerType = {
  individual: 'individual',
  organization: 'organization',
} as const;

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  not_due: 'not_due',
  due: 'due',
  paid: 'paid',
  failed: 'failed',
} as const;

export interface SiteSettings {
  bookingOpeningDate: string;
  earlyBookingMode: boolean;
  pricingVersion?: string;
}

export type BookingInputPreferredContactMethod = typeof BookingInputPreferredContactMethod[keyof typeof BookingInputPreferredContactMethod];


export const BookingInputPreferredContactMethod = {
  whatsapp: 'whatsapp',
  phone: 'phone',
  email: 'email',
} as const;

export interface BookingInput {
  customerType: CustomerType;
  /** @nullable */
  packageId?: string | null;
  /** @nullable */
  pricingVersion?: string | null;
  /**
     * @maxLength 160
     * @nullable
     */
  organizationName?: string | null;
  /**
     * @minLength 2
     * @maxLength 120
     */
  fullName: string;
  /**
     * @minLength 7
     * @maxLength 30
     */
  phone: string;
  /** @nullable */
  email?: string | null;
  preferredContactMethod: BookingInputPreferredContactMethod;
  /**
     * @minLength 2
     * @maxLength 80
     */
  eventType: string;
  eventDate: string;
  /** @pattern ^([01]\d|2[0-3]):[0-5]\d$ */
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  /**
     * @minimum 2
     * @maximum 24
     */
  durationHours: number;
  /**
     * @minLength 2
     * @maxLength 160
     */
  venue: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  location: string;
  /**
     * @minimum 1
     * @maximum 10000
     */
  guestCount: number;
  /**
     * @minLength 2
     * @maxLength 80
     */
  printFormat: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  backdropPreference: string;
  addOns?: string[];
  /**
     * @maxLength 2000
     * @nullable
     */
  brandedRequirements?: string | null;
  /**
     * @maxLength 2000
     * @nullable
     */
  notes?: string | null;
  consent: boolean;
  /**
     * Honeypot field
     * @nullable
     */
  website?: string | null;
  earlyRequest?: boolean;
}

export interface BookingConfirmation {
  id: number;
  reference: string;
  status: BookingStatus;
  potentialConflict: boolean;
  /** @nullable */
  packageName?: string | null;
  quoteRequired?: boolean;
  /** @nullable */
  hourlyRateRwf: number | null;
  /** @nullable */
  totalAmountRwf: number | null;
  /** @nullable */
  depositPercentage: number | null;
  /** @nullable */
  depositAmountRwf: number | null;
  paymentStatus: PaymentStatus;
}

export interface AdminNote {
  id: number;
  note: string;
  createdAt: string;
}

export type BookingNotificationStatus = typeof BookingNotificationStatus[keyof typeof BookingNotificationStatus];


export const BookingNotificationStatus = {
  sent: 'sent',
  pending: 'pending',
  retrying: 'retrying',
  not_configured: 'not_configured',
  none: 'none',
} as const;

export type Booking = BookingInput & ({
  id: number;
  reference: string;
  status: BookingStatus;
  potentialConflict: boolean;
  /** @nullable */
  packageName?: string | null;
  /** @nullable */
  pricingVersion?: string | null;
  /** @nullable */
  basePriceRwf?: number | null;
  /** @nullable */
  includedHours?: number | null;
  quoteRequired?: boolean;
  /** @nullable */
  hourlyRateRwf: number | null;
  /** @nullable */
  totalAmountRwf: number | null;
  /** @nullable */
  depositPercentage: number | null;
  /** @nullable */
  depositAmountRwf: number | null;
  paymentStatus: PaymentStatus;
  /** @nullable */
  paymentMethod: string | null;
  notificationStatus?: BookingNotificationStatus;
  createdAt: string;
  updatedAt: string;
  adminNotes: AdminNote[];
});

export interface BookingUpdate {
  status: BookingStatus;
}

export interface AdminNoteInput {
  /**
     * @minLength 1
     * @maxLength 3000
     */
  note: string;
}

export type AvailabilityWindowReason = typeof AvailabilityWindowReason[keyof typeof AvailabilityWindowReason];


export const AvailabilityWindowReason = {
  confirmed: 'confirmed',
  blocked: 'blocked',
} as const;

export interface AvailabilityWindow {
  date: string;
  /** @nullable */
  startTime: string | null;
  /** @nullable */
  endTime: string | null;
  reason: AvailabilityWindowReason;
  startsAt?: string;
  endsAt?: string;
}

export interface BlockedAvailabilityInput {
  date: string;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endTime?: string | null;
  /**
     * @minLength 2
     * @maxLength 200
     */
  reason: string;
}

export type BlockedAvailability = BlockedAvailabilityInput & {
  id: number;
  createdAt: string;
};

export interface AdminSummary {
  total: number;
  pending: number;
  contacted: number;
  confirmed: number;
  declined: number;
  cancelled: number;
  potentialConflicts: number;
  upcomingConfirmed: number;
}

export type GetAvailabilityParams = {
from: string;
to: string;
};

export type ListBookingsParams = {
status?: BookingStatus;
search?: string;
};