# lib/api-zod/src/generated/types/availabilityWindow.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Memento booking and availability API
 * OpenAPI spec version: 1.0.0
 */
import type { AvailabilityWindowReason } from './availabilityWindowReason';

export interface AvailabilityWindow {
  date: Date;
  /** @nullable */
  startTime: string | null;
  /** @nullable */
  endTime: string | null;
  reason: AvailabilityWindowReason;
  startsAt?: Date;
  endsAt?: Date;
}