# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Memento booking and availability API
 * OpenAPI spec version: 1.0.0
 */
import type { AdminNote } from './adminNote';
import type { BookingInput } from './bookingInput';
import type { BookingNotificationStatus } from './bookingNotificationStatus';
import type { BookingStatus } from './bookingStatus';
import type { PaymentStatus } from './paymentStatus';

export type Booking = BookingInput & ({
  id: number;
  reference: string;
  status: BookingStatus;
  potentialConflict: boolean;
  /** @nullable */
  packageName?: string | null;
  /** @nullable */
  pricingVersion?: string | null;
  /** @nullable */
  basePriceRwf?: number | null;
  /** @nullable */
  includedHours?: number | null;
  quoteRequired?: boolean;
  /** @nullable */
  hourlyRateRwf: number | null;
  /** @nullable */
  totalAmountRwf: number | null;
  /** @nullable */
  depositPercentage: number | null;
  /** @nullable */
  depositAmountRwf: number | null;
  paymentStatus: PaymentStatus;
  /** @nullable */
  paymentMethod: string | null;
  notificationStatus?: BookingNotificationStatus;
  createdAt: Date;
  updatedAt: Date;
  adminNotes: AdminNote[];
});