# lib/api-zod/src/generated/types/bookingConfirmation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Memento booking and availability API
 * OpenAPI spec version: 1.0.0
 */
import type { BookingStatus } from './bookingStatus';
import type { PaymentStatus } from './paymentStatus';

export interface BookingConfirmation {
  id: number;
  reference: string;
  status: BookingStatus;
  potentialConflict: boolean;
  /** @nullable */
  packageName?: string | null;
  quoteRequired?: boolean;
  /** @nullable */
  hourlyRateRwf: number | null;
  /** @nullable */
  totalAmountRwf: number | null;
  /** @nullable */
  depositPercentage: number | null;
  /** @nullable */
  depositAmountRwf: number | null;
  paymentStatus: PaymentStatus;
}