# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Memento booking and availability API
 * OpenAPI spec version: 1.0.0
 */
import type { BookingInputPreferredContactMethod } from './bookingInputPreferredContactMethod';
import type { CustomerType } from './customerType';

export interface BookingInput {
  customerType: CustomerType;
  /** @nullable */
  packageId?: string | null;
  /** @nullable */
  pricingVersion?: string | null;
  /**
     * @maxLength 160
     * @nullable
     */
  organizationName?: string | null;
  /**
     * @minLength 2
     * @maxLength 120
     */
  fullName: string;
  /**
     * @minLength 7
     * @maxLength 30
     */
  phone: string;
  /** @nullable */
  email?: string | null;
  preferredContactMethod: BookingInputPreferredContactMethod;
  /**
     * @minLength 2
     * @maxLength 80
     */
  eventType: string;
  eventDate: Date;
  /** @pattern ^([01]\d|2[0-3]):[0-5]\d$ */
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  /**
     * @minimum 2
     * @maximum 24
     */
  durationHours: number;
  /**
     * @minLength 2
     * @maxLength 160
     */
  venue: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  location: string;
  /**
     * @minimum 1
     * @maximum 10000
     */
  guestCount: number;
  /**
     * @minLength 2
     * @maxLength 80
     */
  printFormat: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  backdropPreference: string;
  addOns?: string[];
  /**
     * @maxLength 2000
     * @nullable
     */
  brandedRequirements?: string | null;
  /**
     * @maxLength 2000
     * @nullable
     */
  notes?: string | null;
  consent: boolean;
  /**
     * Honeypot field
     * @nullable
     */
  website?: string | null;
  earlyRequest?: boolean;
}