# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Memento booking and availability API
 * OpenAPI spec version: 1.0.0
 */

export interface SiteSettings {
  bookingOpeningDate: Date;
  earlyBookingMode: boolean;
  pricingVersion?: string;
}